moved AbstractProvidesOutgoingPartitionConstraints back to the application level.
